% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{analyse}
\alias{analyse}
\alias{analyse.ColliderBias}
\title{Run collider bias analysis}
\usage{
analyse(x, methods, ip, pi0, sxy1, bootstraps)

\method{analyse}{ColliderBias}(
  x,
  methods = NULL,
  ip = NULL,
  pi0 = NULL,
  sxy1 = NULL,
  bootstraps = NULL
)
}
\arguments{
\item{x}{a ColliderBias object}

\item{methods}{a character vector, one or more valid collider bias analysis methods \code{slopehunter}, \code{dudbridge}, \code{ivw_mr}}

\item{ip}{a numeric, range 0-1, the initial p-value by which to filter the incidence GWAS}

\item{pi0}{a numeric, range 0-1, the initial weight / percentage of SNPs that affect incidence only}

\item{sxy1}{a numeric, the initial (guess) of the covariance between incidence and progression BETAs}

\item{bootstraps}{a whole-number numeric, the number of bootstrap samples to estimate the SE of the adjustment factor (can be zero)}
}
\value{
a ColliderBias object with updated parameter and results fields
}
\description{
Run all the analyses defined in the ColliderBias object \code{x} with all combinations
of the listed parameters. Be careful with large numbers of parameters as this will
quickly lead to very large numbers of combinations. e.g. \cr
\itemize{
\item methods = c("slopehunter")
\item ip = c(0.05,0.001,0.00001)
\item pi0 = c(0.6, 0.65, 0.7)
\item sxy1 = c(1e-5, 1e-4, 1e-3)
}
...will lead to 27 separate analyses. \cr
}
