% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{analyse_collider_bias}
\alias{analyse_collider_bias}
\title{Run collider bias analysis}
\usage{
analyse_collider_bias(
  gwas_i,
  gwas_p,
  merge = c(CHR = "CHR", BP = "BP"),
  methods = c("slopehunter", "ivw_mr_collider_bias", "dudbridge"),
  ip = c(0.9),
  pi0 = c(0.6),
  sxy1 = c(1e-05),
  bootstraps = c(100),
  seed = 2023
)
}
\arguments{
\item{gwas_i}{a data.frame like object, clumped incidence GWAS summary statistics in \code{standardise_gwas()} format}

\item{gwas_p}{a data.frame like object, progression GWAS summary statistics in \code{standardise_gwas()} format}

\item{merge}{a named character vector c(gwas1_col=gwas2_col) of the columns to join on; e.g. c("CHR1"="CHR2","BP1"="BP2"); can be NULL, in which case the gwases must be data.tables and already have keys set}

\item{methods}{a character vector of collider bias method function names to run}

\item{ip}{a numeric, range 0-1, the initial p-value by which to filter the incidence GWAS}

\item{pi0}{a numeric, range 0-1, the initial weight / percentage of SNPs that affect incidence only}

\item{sxy1}{a numeric, the initial (guess) of the covariance between incidence and progression BETAs}

\item{bootstraps}{a whole-number numeric, the number of bootstrap samples to estimate the SE of the adjustment factor
(can be zero, in which case no SE is returned)}

\item{seed}{an integer, seed for reproducibility}
}
\value{
a list of S3 ColliderBiasResult objects
}
\description{
Run a set of collider bias analyses with all combinations of input parameters.
Be careful with large numbers of parameters as this will quickly lead to very
large numbers of combinations. e.g. \cr
\itemize{
\item methods = c("slopehunter")
\item ip = c(0.05,0.001,0.00001)
\item pi0 = c(0.6, 0.65, 0.7)
\item sxy1 = c(1e-5, 1e-4, 1e-3)
}
...will lead to 27 separate analyses. \cr
}
