% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{apply_correction}
\alias{apply_correction}
\alias{apply_correction.ColliderBias}
\title{Apply collider bias correction factor}
\usage{
apply_correction(x, b, se)

\method{apply_correction}{ColliderBias}(x, b, se)
}
\arguments{
\item{x}{a ColliderBias object}

\item{b}{the calculated adjustment factor \code{b}}

\item{se}{the calculated b standard error}
}
\value{
a data.table
}
\description{
Apply collider bias adjustment.
}
