% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{clump}
\alias{clump}
\title{Clump a GWAS}
\usage{
clump(
  gwas,
  p1 = 1,
  p2 = 1,
  r2 = 0.1,
  kb = 250,
  plink = genepi.utils::which_plink(),
  plink_ref = genepi.utils::which_1000G_reference()
)
}
\arguments{
\item{gwas}{a data.frame like object with at least columns RSID, EA, OA, and P}

\item{p1}{a numeric, the p-value threshold for inclusion as a clump}

\item{p2}{a numeric, the p-value threshold for incorporation into a clump}

\item{r2}{a numeric, the r2 value}

\item{kb}{a integer, the window for clumping}

\item{plink}{a string, path to the plink executable}

\item{plink_ref}{a string, path to the pfile genome reference}
}
\value{
a data.table with additional columns \code{index} (logical, whether the variant is an index SNP) and \code{clump} (integer, the clump the variant belongs to)
}
\description{
Clump variants in a GWAS using PLINK and an appropriate reference panel.
For example, the 1000 genomes phase 3 data can be downloaded from the PLINK
website (https://www.cog-genomics.org/plink/2.0/resources#phase3_1kg). To remove
duplicates you should run 'plink2 --pfile all_phase3 --rm-dup force-first --make-pgen --out all_phase3_nodup'
The path to the reference (without the plink extensions) should be passed as the
\code{plink_ref} argument. The path to the plink executable should be passed as the
\code{plink} argument.
}
