% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{clump_mr}
\alias{clump_mr}
\title{Clump MR object exposure}
\usage{
clump_mr(
  x,
  p1 = 1,
  p2 = 1,
  r2 = 0.001,
  kb = 250,
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37"),
  parallel_cores = parallel::detectCores()
)
}
\arguments{
\item{x}{an object of class MR description}

\item{p1}{a numeric, the p-value threshold for inclusion as a clump}

\item{p2}{a numeric, the p-value threshold for incorporation into a clump}

\item{r2}{a numeric, the r2 value}

\item{kb}{a integer, the window for clumping}

\item{plink2}{a string, path to the plink executable}

\item{plink_ref}{a string, path to the pfile genome reference}

\item{parallel_cores}{an integer, how many cores / threads to use}
}
\description{
Clump MR object exposure
}
