% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{collider_bias}
\alias{collider_bias}
\title{Run collider bias assessment}
\usage{
collider_bias(
  x,
  bias_method = "dudbridge",
  r2 = 0.001,
  p1 = 5e-08,
  kb = 250,
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37"),
  ip = 0.001,
  pi0 = 0.6,
  sxy1 = 1e-05,
  bootstraps = 100,
  weighted = TRUE,
  method = "Simex",
  B = 1000,
  seed = 2023
)
}
\arguments{
\item{x}{an object of class MR}

\item{bias_method}{a character or character vector, one or more of c("dudbridge", "slopehunter", "mr_ivw", "mr_egger", "mr_weighted_median", "mr_weighted_mode")}

\item{r2}{a numeric 0-1, r2 used for clumping - set all clumping params to NA to turn off}

\item{p1}{a numeric 0-1, p1 used for clumping - set all clumping params to NA to turn off}

\item{kb}{an integer, kb used for clumping - set all clumping params to NA to turn off}

\item{plink2}{a path, the plink2 binary}

\item{plink_ref}{a path, the reference genome pfile}

\item{ip}{a numeric 0-1, threshold for removing incidence variants;  see \code{xp_thresh} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{pi0}{a numeric 0-1, proportion of SNPs in the incidence only cluster; see \code{init_pi} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{sxy1}{a numeric, the covariance between incidence and progression Gip SNPs; see \code{init_sigmaIP} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{bootstraps}{an integer, number of bootstraps to estimate SE; see \code{M} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{weighted}{see \code{weighted} \code{\link[indexevent:indexevent]{indexevent::indexevent()}}}

\item{method}{see \code{method} \code{\link[indexevent:indexevent]{indexevent::indexevent()}}}

\item{B}{see \code{B} \code{\link[indexevent:indexevent]{indexevent::indexevent()}}}

\item{seed}{seed, for reproducibility}
}
\description{
Run collider bias assessment
}
