% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{collider_results_to_data_table}
\alias{collider_results_to_data_table}
\alias{collider_results_to_data_table.ColliderBiasResult}
\alias{collider_results_to_data_table.list}
\title{ColliderBiasResult to data.table}
\usage{
collider_results_to_data_table(results)

\method{collider_results_to_data_table}{ColliderBiasResult}(results)

\method{collider_results_to_data_table}{list}(results)
}
\arguments{
\item{results}{a ColliderBiasResult object, or list of ColliderBiasResult objects}
}
\description{
ColliderBiasResult to data.table
}
