% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{ColliderBias}
\alias{ColliderBias}
\title{ColliderBias class}
\usage{
ColliderBias(
  gwas_i,
  gwas_p,
  methods = c("slopehunter", "dudbridge", "ivw_mr"),
  ip = 0.001,
  pi0 = 0.6,
  sxy1 = 1e-05,
  bootstraps = 100
)
}
\arguments{
\item{gwas_i}{a data.frame like object, incidence GWAS summary statistics in \code{standardise_gwas()} format}

\item{gwas_p}{a data.table like object, progression GWAS summary statistics in \code{standardise_gwas()} format}

\item{methods}{a character vector, one or more valid collider bias analysis methods \code{slopehunter}, \code{dudbridge}, \code{ivw_mr}}

\item{ip}{a numeric, range 0-1, the initial p-value by which to filter the incidence GWAS}

\item{pi0}{a numeric, range 0-1, the initial weight / percentage of SNPs that affect incidence only}

\item{sxy1}{a numeric, the initial (guess) of the covariance between incidence and progression BETAs}

\item{bootstraps}{a whole-number numeric, the number of bootstrap samples to estimate the SE of the adjustment factor (can be zero)}
}
\value{
an S3 object of class ColliderBias
}
\description{
An S3 class for collider bias analysis. This object can be passed around the
collider bias analysis tools, used for plotting, and for applying the chosen
adjustment to your GWAS summary statistics.
}
