% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{ColliderBiasResult}
\alias{ColliderBiasResult}
\title{Collider bias result object}
\usage{
ColliderBiasResult(
  method = NA_character_,
  ip = NA_real_,
  pi0 = NA_real_,
  pi = NA_real_,
  sxy1 = NA_real_,
  b = NA_real_,
  bse = NA_real_,
  entropy = NA_real_,
  fit = data.table::data.table()
)
}
\arguments{
\item{method}{a string, the method used}

\item{ip}{a numeric, the initialising p-value}

\item{pi0}{a numeric, the initialising Pi value for the Slope-hunter method}

\item{pi}{a numeric, the final Pi value, proportion of SNPs in the Gi SNP cluster}

\item{sxy1}{a numeric, the initialising SNPip Cov value for the Slope-hunter method}

\item{b}{a numeric, the estimated correction factor}

\item{bse}{a numeric, the standard error of the correction factor}

\item{entropy}{a numeric, the entropy - output from the Slope-hunter method}

\item{fit}{a data.table, the raw points data for plotting table columns \code{c("SNP_incidence","BETA_incidence","BETA_progression","CLUSTER")}}
}
\value{
an S3 object of class ColliderBiasResult
}
\description{
An S3 class which is the standard output from various collider bias analyses.
}
