% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_colmap.R
\name{ColumnMap}
\alias{ColumnMap}
\title{ColumnMap object}
\usage{
ColumnMap(x)
}
\arguments{
\item{x}{either a list of \code{Column} class objects, a valid string for a pre-defined map: default, metal, ieu_ukb, ieugwasr, ns_map,
gwama, giant, or a named character vector or list (standard name = old name)}
}
\value{
an S7 class genepi.utils::ColumnMap object
}
\description{
A mapping to the standardised column names used in this package. Available names: 'rsid', 'chr', 'bp', 'ea', 'oa', 'eaf', 'p', 'beta',
'se', 'or', 'or_se', 'or_lb', 'or_ub', 'beta_lb', 'beta_ub', 'z', 'q_stat', 'i2', 'nstudies', 'n'
}
\section{Slots}{

\describe{
\item{\code{map}}{a list of \code{Column} class objects}
}}

