% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{cwls}
\alias{cwls}
\title{Corrected Weighted Least Squares collider bias method}
\usage{
cwls(x, ...)
}
\arguments{
\item{x}{an object of class MR}

\item{...}{parameter sink, additional ignored parameters}
}
\value{
an object of class MRResult
}
\description{
Corrected Weighted Least Squares collider bias method
}
