% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{download_ukbb_ld}
\alias{download_ukbb_ld}
\title{Title}
\usage{
download_ukbb_ld(chr, bp_start, bp_end, ukbb_ld_cache)
}
\arguments{
\item{chr}{.}

\item{bp_start}{.}

\item{bp_end}{.}

\item{ukbb_ld_cache}{.}
}
\value{
a data.table of paths to the LD reference files
}
\description{
Title
}
