% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaf_plot.R
\name{eaf_plot}
\alias{eaf_plot}
\title{Effect allele frequency plot}
\usage{
eaf_plot(
  gwas,
  eaf_col = "EAF",
  ref_eaf_col = "EUR_EAF",
  tolerance = 0.2,
  colours = list(missing = "#5B1A18", outlier = "#FD6467", within = "#7294D4"),
  title = NULL,
  facet_grid_row_col = NULL,
  facet_grid_col_col = NULL
)
}
\arguments{
\item{gwas}{a data.table}

\item{eaf_col}{a string, the column containing the study EAF data}

\item{ref_eaf_col}{a string, the column containing the reference EAF data}

\item{tolerance}{a numeric, frequency difference that determines outliers}

\item{colours}{a 3 element list of colour codes, e.g. list(missing="#5B1A18", outlier="#FD6467", within="#7294D4")}

\item{title}{a string, the plot title}

\item{facet_grid_row_col}{(optional), a column by which to facet the plot by rows}

\item{facet_grid_col_col}{(optional), a column by which to facet the plot by columns}
}
\value{
a ggplot
}
\description{
Plotting reported effect allele frequencies (EAF) against a reference set to identify
study variants which significantly deviate from the expected population frequencies.
}
