% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_random_gwas_data}
\alias{generate_random_gwas_data}
\title{Generate random GWAS data}
\usage{
generate_random_gwas_data(n, seed = 2023)
}
\arguments{
\item{n}{number of fake variants to generate}

\item{seed}{seed, for reproducibility}
}
\value{
a data.table with columns SNP, CHR, BP, OA, EA, EAF, BETA, P, EUR_EAF
}
\description{
Generates rows of synthetic GWAS summary stats data. Useful for developing plotting and other
methods. No attempt is made to make this data at all realistic.
}
