% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_column_map.R
\name{get_map}
\alias{get_map}
\alias{get_map.ColumnMapping}
\title{Get column mapping}
\usage{
get_map(x, input_format)

\method{get_map}{ColumnMapping}(x, input_format)
}
\arguments{
\item{x}{a ColumnMapping object}

\item{input_format}{a string, must be a valid input_format}
}
\value{
a named list (standardised_name: original_name)
}
\description{
Get column mapping
}
