% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{get_pfile_variants}
\alias{get_pfile_variants}
\title{Extract variants from plink binary}
\usage{
get_pfile_variants(
  snp,
  win_kb,
  chr,
  from_bp,
  to_bp,
  plink2 = genepi.utils::which_plink2(),
  pfile = genepi.utils::which_1000G_reference(build = "GRCh37")
)
}
\arguments{
\item{snp}{character, an rsid}

\item{win_kb}{numeric, window size around snp in kb}

\item{chr}{character, the chromosome (use instead of snp and win_kb, not in addition)}

\item{from_bp}{numeric, the start base position (use instead of snp and win_kb, not in addition)}

\item{to_bp}{numeric, the end base position (use instead of snp and win_kb, not in addition)}

\item{plink2}{character / path, the plink2 executable}

\item{pfile}{character / path, the plink pfile set}
}
\value{
a data.table
}
\description{
Extract variants from plink binary
}
