% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink.R
\name{get_proxies}
\alias{get_proxies}
\title{Get proxies for variants from plink binary}
\usage{
get_proxies(
  snps,
  stat = "r2-unphased",
  win_kb = 1000,
  win_r2 = 0.2,
  win_ninter = Inf,
  plink2 = genepi.utils::which_plink2(),
  pfile = genepi.utils::which_1000G_reference(build = "GRCh37")
)
}
\arguments{
\item{snps}{character vector, rsids}

\item{stat}{character, the R stat to calculate, one of "r2-unphased", "r2-phased", "r-unphased", "r-phased"}

\item{win_kb}{numeric, the window to look in around the variants}

\item{win_r2}{numeric, the lower r2 limit to include in output, (for --r-phased and --r-unphased, this means |r|≥sqrt(0.2))}

\item{win_ninter}{numeric, controls the maximum number of other variants allowed between variant-pairs in the report. Inf = off.}

\item{plink2}{character / path, the plink2 executable}

\item{pfile}{character / path, the plink pfile set}
}
\value{
a data.table
}
\description{
Get proxies for variants from plink binary
}
