% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas.R
\name{GWAS}
\alias{GWAS}
\title{GWAS object}
\usage{
GWAS(
  dat,
  map = "default",
  drop = FALSE,
  fill = FALSE,
  fill_rsid = FALSE,
  missing_rsid = "fill_CHR:BP",
  parallel_cores = parallel::detectCores(),
  dbsnp_dir = genepi.utils::which_dbsnp_directory(),
  filters = list(beta_invalid = "!is.infinite(beta) & abs(beta) < 20", eaf_invalid =
    "eaf > 0 & eaf < 1", p_invalid = "!is.infinite(p)", se_invalid = "!is.infinite(se)",
    alleles_invalid = "!is.na(ea) & !is.na(oa)", chr_missing = "!is.na(chr)", bp_missing
    = "!is.na(bp)", beta_missing = "!is.na(beta)", se_missing = "!is.na(se)", p_missing =
    "!is.na(p)", eaf_missing = "!is.na(eaf)"),
  reference = NULL,
  ref_map = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{a valid string file path to be read by \code{data.table::fread} or a \code{data.table::data.table} object; the GWAS data source}

\item{map}{a valid input to the \code{ColumnMap} class constructor (a predefined map string id, a named list or character vector, or a ColumMap object)}

\item{drop}{a logical, whether to drop data source columns not in the column \code{map}}

\item{fill}{a logical, whether to add (NAs) missing columns present in the column \code{map} but not present in the data source}

\item{fill_rsid}{either FALSE or a valid argument for the chrpos_to_rsid \code{build} argument, e.g. "b37_dbsnp156"}

\item{missing_rsid}{a string, how to handle missing rsids: one of "fill_CHR:BP", "fill_CHR:BP_OA_EA", "overwrite_CHR:BP",
"overwrite_CHR:BP:OA:EA", "none", or "leave"}

\item{parallel_cores}{an integer, number of cores to used for RSID mapping, default is maximum machine cores}

\item{dbsnp_dir}{path to the dbsnp directory of fst files see chrpos_to_rsid \code{dbsnp_dir} argument}

\item{filters}{a list of named strings, each to be evaluated as an expression to filter the data during the quality control steps (above)}

\item{reference}{a valid string file path to be read by \code{data.table::fread} or a \code{data.table::data.table} object; the reference data}

\item{ref_map}{a valid input to the \code{ColumnMap} class constructor (a predefined map id (a string), a named list or character vector, or a
ColumMap object) defining at least columns \code{rsid} (or \code{chr}, \code{bp}), \code{ea}, \code{oa} and \code{eaf}.}

\item{verbose}{a logical, whether to print details}

\item{...}{variable capture to be passed to the constructor, e.g. individual vectors for the slots, rather that \code{dat}}
}
\value{
an S7 class genepi.utils::GWAS object
}
\description{
A GWAS object is a container for vectors of GWAS data, a correlation matrix, and
meta-data regarding quality control procedures applied at the point of object creation / data import.
}
\section{Slots}{

\describe{
\item{\code{rsid}}{character, variant ID - usually in rs12345 format, however this can be changed with the \code{missing_rsid} argument}

\item{\code{chr}}{character, chromosome identifier}

\item{\code{bp}}{integer, base position}

\item{\code{ea}}{character, effect allele}

\item{\code{oa}}{character, other allele}

\item{\code{eaf}}{numeric, effect allele frequency}

\item{\code{beta}}{numeric, effect size}

\item{\code{se}}{numeric, effect size standard error}

\item{\code{p}}{numeric, p-value}

\item{\code{n}}{integer, total number of samples}

\item{\code{ncase}}{integer, number of cases}

\item{\code{strand}}{character, the strand + or -}

\item{\code{imputed}}{logical, whether imputed}

\item{\code{info}}{numeric, the info score}

\item{\code{q}}{numeric, the Q statistic for meta analysis results}

\item{\code{q_p}}{numeric, the Q statistic P-value}

\item{\code{i2}}{numeric, the I2 statistic}

\item{\code{proxy_rsid}}{character, proxy variant ID}

\item{\code{proxy_chr}}{character, proxy chromosome identifier}

\item{\code{proxy_bp}}{integer, proxy base position}

\item{\code{proxy_ea}}{character, proxy effect allele}

\item{\code{proxy_oa}}{character, proxy other allele}

\item{\code{proxy_eaf}}{numeric, proxy effect allele frequency}

\item{\code{proxy_r2}}{numeric, proxy r2 with rsid}

\item{\code{trait}}{character, the GWAS trait}

\item{\code{id}}{character, the GWAS identifier}

\item{\code{source}}{character, data source; either the file path, or "data.table" if loaded directly}

\item{\code{correlation}}{matrix, a correlation matrix of signed R values between variants}

\item{\code{map}}{\code{ColumnMap}, a mapping of class ColumnMap}

\item{\code{qc}}{list, a named list of filters; name is the filter expression and value is an integer vector of rows that fail the filter}
}}

