% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise}
\alias{harmonise}
\title{Harmonise GWAS effects}
\usage{
harmonise(
  gwas1,
  gwas2,
  gwas1_trait = "incidence",
  gwas2_trait = "progression",
  merge = NULL
)
}
\arguments{
\item{gwas1}{a data.frame like object or file path, GWAS 1}

\item{gwas2}{a data.frame like object or file path, GWAS 2}

\item{gwas1_trait}{a string, suffix to add to gwas1 column names e.g. progression --> columns chr_progression, bp_progression, ..., etc.}

\item{gwas2_trait}{a string, suffix to add to gwas2 column names e.g. progression --> columns chr_progression, bp_progression, ..., etc.}

\item{merge}{a named character vector c(gwas1_col=gwas2_col) of the columns to join on; e.g. c("CHR1"="CHR2","BP1"="BP2"); can be NULL, in which case the gwases must be data.tables and already have keys set}
}
\value{
a data.table, the harmonised dataset
}
\description{
Harmonise effects across two GWAS datasets. The \code{gwas} objects are expected to be in standard format - see \code{standardise_gwas()}.
Corrects strand for non-palindromic SNPs (flip BETAs). Drop palindromic variants. Drop non-matching variant alleles. \cr
This function is based on the same function in the Slopehunter
and TwoSampleMR packages. I have re-written it here to use
the standardised column names and use pure data.table syntax for increased processing speed.
}
\references{
Slopehunter https://github.com/Osmahmoud/SlopeHunter

TwoSampleMR https://github.com/MRCIEU/TwoSampleMR
}
