% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_gwas.R
\name{harmonise_gwas}
\alias{harmonise_gwas}
\title{Harmonise GWAS}
\usage{
harmonise_gwas(gwas, ref, join = "chr:bp", action = 2, ...)
}
\arguments{
\item{gwas}{a GWAS object, data.table, or file path}

\item{ref}{a GWAS object, data.table, or file path}

\item{join}{a character, either 'chr:pos'(default) or 'rsid', the columns to perform the join on}

\item{action}{an integer, 1-, 2-, or 3-}

\item{...}{additional parameters below}

\item{rmap}{a named vector or list, mapping reference input, standard name = old name (active if using data.table or file path inputs)}

\item{gmap}{a named vector or list, mapping gwas input, standard name = old name (active if using data.table or file path inputs)}
}
\value{
a data.table, harmonised GWAS data
}
\description{
Harmonise GWAS
}
