% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_gwas.R
\name{harmonise_gwases}
\alias{harmonise_gwases}
\title{Harmonise multiple GWAS}
\usage{
harmonise_gwases(...)
}
\arguments{
\item{...}{ellipses of gwas data.tables}
}
\value{
list of harmonised gwases
}
\description{
This takes a list of gwases, gets the SNPs in common across all datasets
and arranges them to be in the same order.
}
