% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{harmonise_ld_dat}
\alias{harmonise_ld_dat}
\title{Re-harmonise against LD matrix reference alleles}
\usage{
harmonise_ld_dat(harm, ld_mat, gwas1_trait = "", gwas2_trait = "")
}
\arguments{
\item{harm}{a data.table, output from \code{harmonise()}}

\item{ld_mat}{a matrix, output from \code{ld_matrix()}}

\item{gwas1_trait}{a string, suffix to add to gwas1 column names e.g. progression --> columns CHR_progression, BP_progression, ..., etc.}

\item{gwas2_trait}{a string, suffix to add to gwas2 column names e.g. progression --> columns CHR_progression, BP_progression, ..., etc.}
}
\value{
a two-element \code{list(harm=harm, ld_mat=ld_mat)}
}
\description{
Based on the TwoSampleMR function (see reference)
}
\references{
\href{https://github.com/MRCIEU/TwoSampleMR/blob/cbd03e6ac58a81922248eb1265cb3ee36b9a76ce/R/other_formats.R#L76}{TwoSampleMR::harmonise_ld_dat()}
}
