% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_column_map.R
\name{input_formats}
\alias{input_formats}
\alias{input_formats.ColumnMapping}
\title{Get valid input formats}
\usage{
input_formats(x)

\method{input_formats}{ColumnMapping}(x)
}
\arguments{
\item{x}{a ColumnMapping object}
}
\value{
a character vector of valid input formats
}
\description{
Get valid input formats
}
