% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_column_map.R
\name{is.input_format}
\alias{is.input_format}
\alias{is.input_format.ColumnMapping}
\title{Check if valid input format}
\usage{
is.input_format(x, input_format)

\method{is.input_format}{ColumnMapping}(x, input_format)
}
\arguments{
\item{x}{a ColumnMapping object}

\item{input_format}{a string, must be a valid input_format}
}
\value{
a logical
}
\description{
Check if valid input format
}
