% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{ld_matrix}
\alias{ld_matrix}
\title{Calculate LD matrix}
\usage{
ld_matrix(
  dat,
  colmap = NULL,
  method = "r",
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37"),
  ukbb_ref = NULL,
  force_all = FALSE
)
}
\arguments{
\item{dat}{data.frame like object, or file path, with at least column \code{rsid}; if columns \code{ea},\code{oa},\code{beta},\code{eaf} are provided then the variants
will be return harmonised to the reference panel (effect allele, data = major allele, reference)}

\item{colmap}{a list, mapping to columns list(rsid=?,ea=?,oa=?,beta=?,eaf=?) where ? can be a character vector in the case of harmonised datasets.
Warning - it is assumed that harmonised datasets are indeed harmonised, if not, any unharmonised variants will be
inappropriately removed.}

\item{method}{a string, either \code{r} or \code{r2}}

\item{plink2}{a string, path to the plink executable}

\item{plink_ref}{a string, path to the pfile genome reference}

\item{ukbb_ref}{path to a UKBB reference file}

\item{force_all}{logical, if TRUE assumes all variants will be found in the reference (e.g. non rsid coded chr:pos_ref_alt)}
}
\value{
an LD matrix if only variants provided, else if alleles provided a list(dat=harmonised data, ld_mat=ld_matrix)
}
\description{
Based on the ieugwasr function (see reference)
}
\references{
\href{https://github.com/MRCIEU/ieugwasr/blob/33e4629f4dacd635c68e690bb5648de529c333cc/R/ld_matrix.R#L92C1-L92C16}{ieugwasr::ld_matrix_local()}
}
