% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan plot}
\usage{
manhattan(
  gwas,
  highlight_snps = NULL,
  highlight_win = 100,
  annotate_snps = NULL,
  colours = c("#d9d9d9", "#bfbfbf"),
  highlight_colour = "#e15758",
  highlight_shape = 16,
  highlight_alpha = 1,
  sig_line_1 = 5e-08,
  sig_line_2 = NULL,
  y_limits = c(NULL, NULL),
  title = NULL,
  subtitle = NULL,
  hit_table = FALSE,
  max_table_hits = 10,
  downsample = 0.1,
  downsample_pval = 0.1
)
}
\arguments{
\item{gwas}{a data.table with a minimum of columns SNP, CHR, BP, and P}

\item{highlight_snps}{(optional) a character vector of SNPs to highlight}

\item{highlight_win}{(optional) a numeric, the number of kb either side of the highlight_snps to also highlight (i.e create peaks)}

\item{annotate_snps}{(optional) a character vector of SNPs to annotate}

\item{colours}{(optional) a character vector colour codes to be replicated along the chromosomes}

\item{highlight_colour}{(optional) a character colour code; the colour to highlight points in}

\item{highlight_shape}{(optional) a numeric shape code; the shape of the highlight points (see ggplot2 shape codes)}

\item{highlight_alpha}{(optional) a numeric value between 0 and 1; the alpha of the highlighted points colour}

\item{sig_line_1}{(optional) a numeric value (-log10(P)) for where to draw a horizontal line}

\item{sig_line_2}{(optional) a numeric value (-log10(P)) for where to draw a second horizontal line}

\item{y_limits}{(optional) a numeric length 2 vector c(min-Y, max-Y)}

\item{title}{(optional) a string title}

\item{subtitle}{(optional) a string subtitle}

\item{hit_table}{(optional) a logical, whether to display a table of top hits (lowest P values)}

\item{max_table_hits}{(optional) an integer, how many top hits to show in the table}

\item{downsample}{(optional) a numeric between 0 and 1, tghe proportion by which to downsample by, e.g. 0.5 will remove 50\% of points above the downsample_pval threshold (can help increase plotting speed with minimal impact on plot appearance)}

\item{downsample_pval}{(optional) a numeric between 0 and 1, the p-values affected by downsampling, default >0.1}
}
\value{
a ggplot
}
\description{
Create a Manhattan plot with ggplot2 geom_point.
}
