% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{miami}
\alias{miami}
\title{Miami plot}
\usage{
miami(
  gwases,
  highlight_snps = list(top = NULL, bottom = NULL),
  highlight_win = list(top = 100, bottom = 100),
  annotate_snps = list(top = NULL, bottom = NULL),
  colours = list(top = c("#d9d9d9", "#bfbfbf"), bottom = c("#bfbfbf", "#d9d9d9")),
  highlight_colour = list(top = "#e15758", bottom = "#4f79a7"),
  highlight_shape = list(top = 16, bottom = 16),
  sig_line_1 = list(top = 5e-08, bottom = 5e-08),
  sig_line_2 = list(top = NULL, bottom = NULL),
  y_limits = list(top = c(NULL, NULL), bottom = c(NULL, NULL)),
  title = NULL,
  subtitle = list(top = NULL, bottom = NULL),
  hit_table = FALSE,
  max_table_hits = 10,
  downsample = 0.1,
  downsample_pval = 0.1
)
}
\arguments{
\item{gwases}{a list of 2 data.tables}

\item{highlight_snps}{(optional) a character vector of SNPs to highlight}

\item{highlight_win}{(optional) a numeric, the number of kb either side of the highlight_snps to also highlight (i.e create peaks)}

\item{annotate_snps}{(optional) a character vector of SNPs to annotate}

\item{colours}{(optional) a character vector colour codes to be replicated along the chromosomes}

\item{highlight_colour}{(optional) a character colour code; the colour to highlight points in}

\item{highlight_shape}{(optional) a numeric shape code; the shape of the highlight points (see ggplot2 shape codes)}

\item{sig_line_1}{(optional) a numeric value (-log10(P)) for where to draw a horizontal line}

\item{sig_line_2}{(optional) a numeric value (-log10(P)) for where to draw a second horizontal line}

\item{y_limits}{(optional) a numeric length 2 vector c(min-Y, max-Y)}

\item{title}{(optional) a string title}

\item{subtitle}{(optional) a string subtitle}

\item{hit_table}{(optional) a logical, whether to display a table of top hits (lowest P values)}

\item{max_table_hits}{(optional) an integer, how many top hits to show in the table}

\item{downsample}{(optional) a numeric between 0 and 1, tghe proportion by which to downsample by, e.g. 0.5 will remove 50\% of points above the downsample_pval threshold (can help increase plotting speed with minimal impact on plot appearance)}

\item{downsample_pval}{(optional) a numeric between 0 and 1, the p-values affected by downsampling, default >0.1}
}
\value{
a ggplot
}
\description{
Create a Miami plot. Please look carefully at the parameters as these largely map to the
\code{manhattan()} parameters, the main difference being that you need to supply a 2 element
list of the parameter, one for the upper and one for the lower plot aspect of the Miami
plot. Some parameters are not duplicated however - see the example defaults below.
}
