% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{MR}
\alias{MR}
\title{MR object}
\usage{
MR(
  exposure,
  outcome,
  harmonise_strictness = 2,
  plink2 = genepi.utils::which_plink2(),
  pfile = genepi.utils::which_1000G_reference(build = "GRCh37"),
  correlation = NULL,
  verbose = TRUE
)
}
\arguments{
\item{exposure}{a \code{GWAS} object or list of \code{GWAS} objects}

\item{outcome}{a \code{GWAS} object}

\item{harmonise_strictness}{an integer (1,2,3) corresponding to the TwoSampleMR harmonisation options of the same name.}

\item{plink2}{character / path, the plink2 executable}

\item{pfile}{character / path, the plink pfile set}

\item{correlation}{a matrix, correlation matrix of signed R values between variants}

\item{verbose}{a logical, print more information}
}
\value{
an S7 class genepi.utils::MR object
}
\description{
An MR object is a container for vectors and matrices of 2 or more GWAS data. \cr\cr
}
\section{Slots}{

\describe{
\item{\code{snps}}{character, variant ID}

\item{\code{chr}}{character, chromosome identifier}

\item{\code{bp}}{integer, base position}

\item{\code{ea}}{character, effect allele}

\item{\code{oa}}{character, other allele}

\item{\code{eafx}}{numeric, exposure effect allele frequency}

\item{\code{nx}}{integer, exposure total number of samples}

\item{\code{ncasex}}{integer, exposure number of cases}

\item{\code{bx}}{numeric, exposure effect size}

\item{\code{bxse}}{numeric, exposure effect size standard error}

\item{\code{px}}{numeric, exposure p-value}

\item{\code{eafy}}{numeric, exposure effect allele frequency}

\item{\code{ny}}{integer, exposure total number of samples}

\item{\code{ncasey}}{integer, exposure number of cases}

\item{\code{by}}{numeric, exposure effect size}

\item{\code{byse}}{numeric, exposure effect size standard error}

\item{\code{py}}{numeric, exposure p-value}

\item{\code{exposure_id}}{character, the GWAS identifier}

\item{\code{exposure}}{character, the GWAS exposure}

\item{\code{outcome_id}}{character, the GWAS identifier}

\item{\code{outcome}}{character, the GWAS outcome}

\item{\code{group}}{integer, grouping variable used for plotting}

\item{\code{index_snp}}{logical, whether the variant is an index variant (via clumping)}

\item{\code{proxy_snp}}{character, the id of the proxy snp}

\item{\code{ld_info}}{logical, whether there is LD information}

\item{\code{info}}{data.frame, information about the loaded GWAS objects}

\item{\code{correlation}}{matrix, a correlation matrix of signed R values between variants}
}}

