% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{mr_collider_bias}
\alias{mr_collider_bias}
\title{IVW MR collider bias method}
\usage{
mr_collider_bias(
  gwas_i,
  gwas_p,
  ip = 0.001,
  merge = c(CHR = "CHR", BP = "BP"),
  tsmr_method = "mr_ivw",
  ...
)
}
\arguments{
\item{gwas_i}{a data.frame like object, clumped incidence GWAS summary statistics in \code{standardise_gwas()} format}

\item{gwas_p}{a data.frame like object, progression GWAS summary statistics in \code{standardise_gwas()} format}

\item{ip}{a numeric, range 0-1, the initial p-value by which to filter the incidence GWAS}

\item{merge}{a named character vector c(gwas1_col=gwas2_col) of the columns to join on; e.g. c("CHR1"="CHR2","BP1"="BP2"); can be NULL, in which case the gwases must be data.tables and already have keys set}

\item{tsmr_method}{a string; a valid \code{TwoSampleMR::mr()} \code{method_list} parameter}

\item{...}{parameter sink, additional ignored parameters}
}
\value{
an S3 ColliderBiasResult object
}
\description{
The Inverse variance weighted Mendelian Randomisation method for assessing collider bias.
}
