% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{mr_egger}
\alias{mr_egger}
\title{Run Egger MR}
\usage{
mr_egger(x, corr = FALSE, ...)
}
\arguments{
\item{x}{an object of class MR}

\item{corr}{a logical, whether to use the correlation matrix when running MR}

\item{...}{parameter sink, not used}
}
\description{
Run Egger MR
}
