% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{mr_ivw}
\alias{mr_ivw}
\title{Run IVW MR}
\usage{
mr_ivw(x, corr = FALSE, ...)
}
\arguments{
\item{x}{an object of class MR}

\item{corr}{a logical, whether to use the correlation matrix when running MR}

\item{...}{parameter sink, not used}
}
\description{
Run IVW MR
}
