% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{mr_pcgmm}
\alias{mr_pcgmm}
\title{Run PC-GMM MR}
\usage{
mr_pcgmm(x, corr = TRUE, ...)
}
\arguments{
\item{x}{an object of class MR}

\item{corr}{a logical, whether to use the correlation matrix when running MR}

\item{...}{parameter sink, not used}
}
\description{
Run PC-GMM MR
}
