% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{mr_results_to_data_table}
\alias{mr_results_to_data_table}
\title{MR results to data.table}
\usage{
mr_results_to_data_table(x)
}
\arguments{
\item{x}{MRResult object to covert to data.table}
}
\description{
MR results to data.table
}
