% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{plot_correction_stability}
\alias{plot_correction_stability}
\title{Plot the estimated correction factor variability}
\usage{
plot_correction_stability(results)
}
\arguments{
\item{results}{a list of ColliderBiasResult objects, from \code{analyse_collider_bias()}}
}
\value{
a ggplot2 plot
}
\description{
Plot the estimated correction factor variability
}
