% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_target_proxy.R
\name{plot_drug_proxy_instrument}
\alias{plot_drug_proxy_instrument}
\title{Plot the drug proxy instrument}
\usage{
plot_drug_proxy_instrument(dat, remove = c("gwas_pthresh", "clumping"))
}
\arguments{
\item{dat}{a data.table, output from `drug_target_proxy()``}

\item{remove}{a character vector, one or more of c("gwas_pthresh", "clumping")}
}
\value{
a ggplot
}
\description{
Plot the drug proxy instrument
}
