% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{plot_mr}
\alias{plot_mr}
\title{Plot MR results}
\usage{
plot_mr(mr, res)
}
\arguments{
\item{mr}{an object of class MR}

\item{res}{a data.table output from run_mr or other MR methods}
}
\description{
Plot MR results
}
