% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{plot_p_threshold}
\alias{plot_p_threshold}
\alias{plot_p_threshold.ColliderBias}
\title{Plot the estimated correction factor variability}
\usage{
plot_p_threshold(x)

\method{plot_p_threshold}{ColliderBias}(x)
}
\arguments{
\item{x}{a ColliderBias object with populated results field}
}
\value{
a plot
}
\description{
Plot the estimated correction factor variability
}
