% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{plot_slope}
\alias{plot_slope}
\title{Plot the Slope-hunter scatter plot}
\usage{
plot_slope(results)
}
\arguments{
\item{results}{a ColliderBiasResult object, or list of ColliderBiasResult objects, from \code{analyse_collider_bias()},
\code{slopehunter()}, \code{ivw_mr_collider_bias()}, or \code{dudbridge()}}
}
\value{
a ggplot2
}
\description{
Plot the Slope-hunter scatter plot
}
