% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{plot_slopehunter}
\alias{plot_slopehunter}
\alias{plot_slopehunter.ColliderBias}
\title{Plot the Slope-hunter scatter plot}
\usage{
plot_slopehunter(x)

\method{plot_slopehunter}{ColliderBias}(x)
}
\arguments{
\item{x}{a ColliderBias object with populated results field}
}
\value{
a plot
}
\description{
Plot the Slope-hunter scatter plot
}
