% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{plot_slopehunter_iters}
\alias{plot_slopehunter_iters}
\alias{plot_slopehunter_iters.ColliderBias}
\title{Create an animation of Slope-hunter iterations}
\usage{
plot_slopehunter_iters(x)

\method{plot_slopehunter_iters}{ColliderBias}(x)
}
\arguments{
\item{x}{a ColliderBias object}
}
\value{
a plot
}
\description{
Create an animation of Slope-hunter iterations
}
