% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_target_proxy.R
\name{QTL}
\alias{QTL}
\title{QTL class}
\usage{
QTL(dat, p_val, join_key = NULL)
}
\arguments{
\item{dat}{a data.table or file path to a data set containing at least columns ...}

\item{p_val}{a numeric, p-value threshold to filter on}

\item{join_key}{a string, column name to join on to other datasets}
}
\value{
an S3 QTL class data structure
}
\description{
A short description...
}
