% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{run_mr}
\alias{run_mr}
\title{Run MR}
\usage{
run_mr(
  x,
  corr = FALSE,
  methods = c("mr_ivw", "mr_egger", "mr_weighted_median", "mr_weighted_mode"),
  ...
)
}
\arguments{
\item{x}{an object of class MR}

\item{corr}{a logical, whether to use the correlation matrix when running MR}

\item{methods}{a string, one of c('mr_ivw','mr_egger','mr_weighted_median','mr_weighted_mode', 'mr_pcgmm')}

\item{...}{parameter sink, not used}
}
\description{
Run MR
}
