% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_1000G_reference}
\alias{set_1000G_reference}
\title{Set the 1000G reference path}
\usage{
set_1000G_reference(path, build = "GRCh37")
}
\arguments{
\item{path}{path to the 1000G reference pfile}

\item{build}{one of c("GRCh37", "GRCh38")}
}
\value{
NULL, updated config file
}
\description{
Set the 1000G reference path
}
