% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dbsnp_directory}
\alias{set_dbsnp_directory}
\title{Set dbSNP directory}
\usage{
set_dbsnp_directory(path)
}
\arguments{
\item{path}{path to the \code{dbsnp} directory}
}
\value{
NULL, updated config file
}
\description{
Set dbSNP directory
}
