% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{set_ld_mat}
\alias{set_ld_mat}
\title{Set the LD matrix}
\usage{
set_ld_mat(x, correlation)
}
\arguments{
\item{x}{an object of class MR}

\item{correlation}{a matrix, the correlation ('r') matrix}
}
\description{
Set the LD matrix
}
