% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_plink2}
\alias{set_plink2}
\title{Set the PLINK2 path}
\usage{
set_plink2(path)
}
\arguments{
\item{path}{path to the PLINK2 executable}
}
\value{
NULL, updated config file
}
\description{
Set the PLINK2 path
}
