% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{slopehunter}
\alias{slopehunter}
\title{Slope-Hunter collider bias method}
\usage{
slopehunter(
  x,
  ip = 0.001,
  pi0 = 0.6,
  sxy1 = 1e-05,
  bootstraps = 100,
  seed = 777
)
}
\arguments{
\item{x}{an object of class MR}

\item{ip}{see \code{xp_thresh} SlopeHunter::hunt()}

\item{pi0}{see \code{init_pi} SlopeHunter::hunt()}

\item{sxy1}{see \code{init_sigmaIP} SlopeHunter::hunt()}

\item{bootstraps}{see \code{M} SlopeHunter::hunt()}

\item{seed}{see \code{seed} SlopeHunter::hunt()}
}
\value{
an object of class MRResult
}
\description{
Slope-Hunter collider bias method
}
