% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_gwas.R
\name{standardise_gwas}
\alias{standardise_gwas}
\title{standardise_gwas}
\usage{
standardise_gwas(gwas, input_format = "default", populate_rsid = FALSE)
}
\arguments{
\item{gwas}{a string file path, or data.frame like object}

\item{input_format}{a string or list()}

\item{populate_rsid}{a logical}
}
\value{
a standardised data.table
}
\description{
Standardises a GWAS with appropriate column names and data checks.
}
