% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gwas.R
\name{subset_gwas}
\alias{subset_gwas}
\title{subset_gwas}
\usage{
subset_gwas(x, snps)
}
\arguments{
\item{x}{GWAS object}

\item{snps}{a vector, either row indicies (integers) into the GWAS object (e.g. obtained
with filters such as which(GWAS'at'p < 5e-8), or rsids (characters) to be found in the
GWAS rsid slot.}
}
\value{
GWAS object subsetted by \code{snps}
}
\description{
subset_gwas
}
