% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{to_MRInput}
\alias{to_MRInput}
\title{Convert to MendelianRandomization::MRInput object}
\usage{
to_MRInput(x, corr = FALSE)
}
\arguments{
\item{x}{an object of class MR}

\item{corr}{a logical, whether to use the correlation matrix when running MR}
}
\description{
Convert to MendelianRandomization::MRInput object
}
