% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_dbsnp_builds}
\alias{which_dbsnp_builds}
\title{Get available dbSNP builds}
\usage{
which_dbsnp_builds(build = NULL)
}
\arguments{
\item{build}{a dbSNP build}
}
\value{
a list of available dbSNP builds - name(dbSNP build): value(directory_path)
}
\description{
Get available dbSNP builds
}
