% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrpos_to_rsid.R
\name{chrpos_to_rsid}
\alias{chrpos_to_rsid}
\title{chrpos_to_rsid}
\usage{
chrpos_to_rsid(
  dt,
  chr_col,
  pos_col,
  ea_col = NULL,
  nea_col = NULL,
  build = "GRCh37",
  flip = "report"
)
}
\arguments{
\item{dt}{a data.table with at least columns (chrom\if{html}{\out{<chr>}}, pos\if{html}{\out{<int>}}, ea\if{html}{\out{<chr>}}, nea\if{html}{\out{<chr>}})}

\item{chr_col}{a string column name; chromosome position}

\item{pos_col}{a string column name; base position}

\item{ea_col}{a string column name; effect allele}

\item{nea_col}{a string column name; non effect allele}

\item{build}{a string, options: "GRCh37"}

\item{flip}{a string, options: "report", "allow", "no_flip"}
}
\value{
a data.table with an RSID column
}
\description{
chrpos_to_rsid
}
