% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{clump_mr}
\alias{clump_mr}
\title{Clump MR object exposure}
\usage{
clump_mr(
  x,
  p1 = 1,
  p2 = 1,
  r2 = 0.001,
  kb = 250,
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37")
)
}
\description{
Clump MR object exposure
}
