% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_colmap.R
\name{Column}
\alias{Column}
\title{Column object}
\usage{
Column(name = character(0), alias = character(0), type = character(0))
}
\arguments{
\item{name}{the standard column name}

\item{alias}{a character vector of aliases (other column names) for this column}

\item{type}{a character, an atomic R type}
}
\value{
an S7 class genepi.utils::Column object
}
\description{
Column object
}
\section{Slots}{

\describe{
\item{\code{name}}{the standard column name}

\item{\code{alias}}{a character vector of aliases (other column names) for this column}

\item{\code{type}}{a character, an atomic R type}
}}

