% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{dudbridge}
\alias{dudbridge}
\title{Dudbridge collider bias method}
\usage{
dudbridge(
  x,
  weighted = TRUE,
  prune = NULL,
  method = "Simex",
  B = 1000,
  lambda = seq(0.25, 5, 0.25),
  seed = 2018,
  ...
)
}
\arguments{
\item{...}{parameter sink, additional ignored parameters}
}
\value{
an object of class MRResult
}
\description{
Dudbridge collider bias method
}
