% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{dudbridge}
\alias{dudbridge}
\alias{dudbridge.ColliderBias}
\title{Dudbridge CWLS method}
\usage{
dudbridge(x)

\method{dudbridge}{ColliderBias}(x)
}
\arguments{
\item{x}{a ColliderBias object}
}
\value{
a ColliderBias object with updated parameter and results fields
}
\description{
The Dudbridge corrected weighted least-squares method for assessing collider bias.
}
