% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepi.utils-package.R
\docType{package}
\name{genepi.utils-package}
\alias{genepi.utils}
\alias{genepi.utils-package}
\title{genepi.utils: GenEpi Utility Functions}
\description{
The `genepi.utils` package is a collection of utility functions for working with genetic epidemiology data.
}
\author{
\strong{Maintainer}: Nicholas Sunderland \email{nicholas.sunderland@bristol.ac.uk} (\href{https://orcid.org/0000-0003-3625-630X}{ORCID})

Authors:
\itemize{
  \item Andrew Elmore
}

}
\keyword{internal}
