% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise.R
\name{harmonise}
\alias{harmonise}
\title{Harmonise GWAS effects}
\usage{
harmonise(gwas1, gwas2, gwas1_trait = "incidence", gwas2_trait = "progression")
}
\arguments{
\item{gwas1}{a data.frame like object, GWAS 1}

\item{gwas2}{a data.frame like object, GWAS 2}
}
\value{
a data.table, the harmonised dataset
}
\description{
Harmonise effects across two GWAS datasets. The \code{gwas} objects are expected to be in standard format - see \code{standardise_gwas()}.
Corrects strand for non-palindromic SNPs (flip BETAs). Drop palindromic variants. Drop non-matching variant alleles. \cr
This function is essentially the same as that of the \href{https://github.com/Osmahmoud/SlopeHunter/blob/master/R/harmonise_effects.R}{Slopehunter}
package, which is turn is based on a similar procedure from the \href{https://github.com/MRCIEU/TwoSampleMR/blob/master/R/harmonise.R}{TwoSampleMR}
package. I have re-written it here to use the standardised column names and use pure data.table syntax. Please credit these other
packages if using this function.
}
\references{
\href{https://github.com/Osmahmoud/SlopeHunter/blob/master/R/harmonise_effects.R}{Slopehunter} https://github.com/Osmahmoud/SlopeHunter/blob/master/R/harmonise_effects.R

\href{https://github.com/MRCIEU/TwoSampleMR/blob/master/R/harmonise.R}{TwoSampleMR} https://github.com/MRCIEU/TwoSampleMR/blob/master/R/harmonise.R
}
