% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{ld_matrix}
\alias{ld_matrix}
\title{Calculate LD matrix}
\usage{
ld_matrix(
  variants,
  with_alleles = TRUE,
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37")
)
}
\arguments{
\item{variants}{data.frame like object, or file path, with at least column \code{RSID}; if columns \code{EA},\code{OA},\code{BETA},\code{EAF} are provided then the variants
will be return harmonised to the reference panel (effect allele, data = major allele, reference)}

\item{plink2}{a string, path to the plink executable}

\item{plink_ref}{a string, path to the pfile genome reference}
}
\value{
an LD matrix if only variants provided, else if alleles provided a list(dat=harmonised data, ld_mat=ld_matrix)
}
\description{
Based on the ieugwasr function (see reference)
}
\references{
\href{https://github.com/MRCIEU/ieugwasr/blob/33e4629f4dacd635c68e690bb5648de529c333cc/R/ld_matrix.R#L92C1-L92C16}{ieugwasr::ld_matrix_local()}
}
