% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_matrix.R
\name{ld_matrix}
\alias{ld_matrix}
\title{Calculate LD matrix}
\usage{
ld_matrix(
  variants,
  plink2 = genepi.utils::which_plink2(),
  plink_ref = genepi.utils::which_1000G_reference(build = "GRCh37"),
  logging = TRUE
)
}
\arguments{
\item{variants}{data.frame like object, or file path, with at least column \code{RSID}}

\item{plink2}{a string, path to the plink executable}

\item{plink_ref}{a string, path to the pfile genome reference}

\item{logging}{a logical, whether to set the plink logging information as attributes (\code{log}, \code{missing_id}, \code{missing_allele}) on the returned data.table}
}
\value{
an LD matrix
}
\description{
Calculate LD matrix
}
