% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift}
\alias{lift}
\title{Liftover GWAS positions}
\usage{
lift(
  gwas,
  from = "Hg19",
  to = "Hg38",
  snp_col = "snp",
  chr_col = "chr",
  pos_col = "pos",
  ea_col = "ea",
  oa_col = "oa"
)
}
\arguments{
\item{gwas}{a data.table, or file path, chr, pos, snp name, effect allele, non-effect allele columns}

\item{to}{which build to lift over to, one of c("Hg18", "Hg19", "Hg38")}

\item{snp_col}{Name of SNP column name. Optional. Uses less certain method of matching if not available}

\item{chr_col}{Name of chromosome column name. Required}

\item{pos_col}{Name of position column name. Required}

\item{ea_col}{Name of effect allele column name. Optional. Might lead to duplicated rows if not presented}

\item{oa_col}{Name of other allele column name. Optional. Might lead to duplicated rows if not presented}

\item{build}{The possible builds, one of c("Hg18", "Hg19", "Hg38")}

\item{build_fallback}{Whether to try "position" (fast) or "biomart" (more accurate if you have rsids) based approaches instead}
}
\value{
data.table with updated position columns
}
\description{
Determine GWAS build and liftover to required build. This is the same function from the
GwasDataImport package, the only difference being that you can specify the build rather
than it trying to guess the build (which fails if you are trying to liftover
small segments of the genome).
}
\references{
https://github.com/MRCIEU/GwasDataImport
}
