% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mr.R
\name{MR}
\alias{MR}
\title{MR object}
\usage{
MR(
  exposure,
  outcome,
  harmonise_strictness = 2,
  with_proxies = FALSE,
  proxy_r2 = 0.8,
  proxy_kb = 250,
  proxy_window = 15,
  proxy_eaf = 0.01,
  correlation = NULL,
  verbose = TRUE
)
}
\arguments{
\item{exposure}{a \code{GWAS} object or list of \code{GWAS} objects}

\item{outcome}{a \code{GWAS} object}

\item{harmonise_strictness}{an integer (1,2,3) corresponding to the TwoSampleMR harmonisation options of the same name.}

\item{with_proxies}{a logical, whether to search for outcome proxy SNPs if the original variant is not present in the data.}

\item{proxy_r2}{a numeric (0-1), the minimum r2 threshold used to define a suitable proxy variant.}

\item{proxy_kb}{an integer, the maximum distance in kb that a proxy variant can be from the index variant.}

\item{proxy_window}{an integer, the maximum number of variants to search either side of the index variant.}

\item{proxy_eaf}{a numeric, the minimum minor allele frequency required to define a suitable proxy.}

\item{correlation}{a matrix, correlation matrix of signed R values between variants}

\item{verbose}{a logical, print more information}
}
\value{
an S7 class genepi.utils::MR object
}
\description{
An MR object is a container for vectors and matrices of 2 or more GWAS data. \cr\cr
}
\section{Slots}{

\describe{
\item{\code{snps}}{character, variant ID}

\item{\code{chr}}{character, chromosome identifier}

\item{\code{bp}}{integer, base position}

\item{\code{ea}}{character, effect allele}

\item{\code{oa}}{character, other allele}

\item{\code{eafx}}{numeric, exposure effect allele frequency}

\item{\code{nx}}{integer, exposure total number of samples}

\item{\code{ncasex}}{integer, exposure number of cases}

\item{\code{bx}}{numeric, exposure effect size}

\item{\code{bxse}}{numeric, exposure effect size standard error}

\item{\code{px}}{numeric, exposure p-value}

\item{\code{eafy}}{numeric, exposure effect allele frequency}

\item{\code{ny}}{integer, exposure total number of samples}

\item{\code{ncasey}}{integer, exposure number of cases}

\item{\code{by}}{numeric, exposure effect size}

\item{\code{byse}}{numeric, exposure effect size standard error}

\item{\code{py}}{numeric, exposure p-value}

\item{\code{exposure_id}}{character, the GWAS identifier}

\item{\code{exposure}}{character, the GWAS exposure}

\item{\code{outcome_id}}{character, the GWAS identifier}

\item{\code{outcome}}{character, the GWAS outcome}

\item{\code{group}}{integer, grouping variable used for plotting}

\item{\code{index_snp}}{logical, whether the variant is an index variant (via clumping)}

\item{\code{proxy_snp}}{character, the id of the proxy snp}

\item{\code{ld_info}}{logical, whether there is LD information}

\item{\code{correlation}}{matrix, a correlation matrix of signed R values between variants}
}}

