% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{slopehunter}
\alias{slopehunter}
\title{Slope-Hunter collider bias method}
\usage{
slopehunter(
  x,
  ip = 0.001,
  pi0 = 0.6,
  sxy1 = 1e-05,
  bootstraps = 100,
  seed = 777,
  ...
)
}
\arguments{
\item{...}{parameter sink, additional ignored parameters}
}
\value{
an object of class MRResult
}
\description{
Slope-Hunter collider bias method
}
