% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collider_bias.R
\name{slopehunter}
\alias{slopehunter}
\title{Slope-Hunter collider bias method}
\usage{
slopehunter(
  x,
  ip = 0.001,
  pi0 = 0.6,
  sxy1 = 1e-05,
  bootstraps = 100,
  seed = 777,
  ...
)
}
\arguments{
\item{x}{an object of class MR}

\item{ip}{see \code{xp_thresh} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{pi0}{see \code{init_pi} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{sxy1}{see \code{init_sigmaIP} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{bootstraps}{see \code{M} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{seed}{see \code{seed} \code{\link[SlopeHunter:hunt]{SlopeHunter::hunt()}}}

\item{...}{parameter sink, additional ignored parameters}
}
\value{
an object of class MRResult
}
\description{
Slope-Hunter collider bias method
}
