% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_gwas.R
\name{standardise_gwas}
\alias{standardise_gwas}
\title{Standardise GWAS format}
\usage{
standardise_gwas(
  gwas,
  input_format = "default",
  build = "GRCh37",
  populate_rsid = FALSE
)
}
\arguments{
\item{gwas}{a string file path, or data.frame like object}

\item{input_format}{a string or list()}

\item{build}{a string, one of GRCh37 or GRCh38}

\item{populate_rsid}{either FALSE or a valid argument for chrpos_to_rsid \code{build}, e.g. b37_dbsnp156}
}
\value{
a standardised data.table
}
\description{
Standardises a GWAS with appropriate column names and data checks.
}
