% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_dbsnp_directory}
\alias{which_dbsnp_directory}
\title{Get dbSNP directory}
\usage{
which_dbsnp_directory()
}
\value{
a string file path, the currently set dbSNP directory path
}
\description{
Get dbSNP directory
}
